// src/temp/attendance/dto/attendance.type.ts
import { Teacher, Person, School, Event, AttendanceEnum, GroupStaff } from "prisma/generated/client";

export interface IAttendanceGrouped {
  personRoleId: number;
  fullName: string;
  totalEvents: number;
}

export interface IMentorPrincipalSchool {
  School: Pick<School, "code" | "name" | "coordenates"> & {
    District: {
      name: string;
      Municipality: {
        name: string;
      };
    };
  };
}

export interface TeacherByUser extends Pick<Teacher, "deletedBy"> {
  Person: Pick<Person, "id" | "firstName" | "lastName1" | "lastName2" | "deletedBy">;
  School: IMentorPrincipalSchool[];
}
export interface ITeachersAssignmentMentor {
  Teacher: TeacherByUser;
}

export interface ITeachersAssignmentMentorResult {
  id: number;
  fullName: string;
  School: Pick<School, "code" | "name" | "coordenates"> & { location: string };
}

export interface IMentors {
  Mentor: {
    id: number;
    deletedBy: number | null;
    firstName: string;
    lastName1: string;
    lastName2: string | null;
  };
}

export interface IMentorResult {
  id: number;
  fullName: string;
}

export interface GetAllSupportResponse {
  Person: Pick<Person, "id" | "firstName" | "lastName1" | "lastName2">;
}

export interface GetAllSupportResult {
  id: number;
  fullName: string;
}

export interface GetAllEventResponse {
  id: number;
  Event: Pick<Event, "name">;
}

export interface GetAllEventResult {
  id: number;
  name: string;
}

export interface GetAllInscriptionResponse {
  Person: Pick<Person, "id" | "firstName" | "lastName1" | "lastName2" | "deletedAt">;
  School: Pick<School, "code" | "name" | "coordenates"> & {
    District: { name: string; Municipality: { name: string } };
  };
}

export interface GetAllInscriptionResult extends GetAllSupportResult {
  School: Pick<School, "code" | "name">;
}

export interface CreateAttendanceSessionData {
  eventInstanceId: number;
  modality: string;
  personId: number;
  trainingModuleId: number;
  coordenates?: string;
}

export interface CreateEventAttendanceData {
  attendanceSessionId: number;
  personId: number;
  status: AttendanceEnum;
}

export interface CreateAttendaceExceptionData {
  eventAttendanceId: number;
  absenceClassificationId: number;
  comment?: string | null;
  justificationFileUrl?: string;
}

export interface FindLastEventResponse {
  id: number;
  checkIn: Date;
  Person: {
    id: number;
    firstName: string;
    lastName1: string;
    lastName2: string | null;
  };
  status: "PRESENTE" | "AUSENTE";
}

export interface FindLastAttendanceResult {
  id: number;
  coordenates: string | null;
  checkIn: Date;
  modality: string;
  eventInstance: {
    id: number;
    name: string;
    responsible: string;
  };
  support: GetAllSupportResult;
  teacherSession: {
    id: number;
    checkIn: Date;
    status: "PRESENTE" | "AUSENTE";
    teacher: {
      id: number;
      fullName: string;
    };
  }[];
}

export interface AttendanceSessionList {
  id: number;
  event: string;
  checkIn: Date;
  checkOut: Date | null;
  modality: string;
  coordenates: string | null;
  support: {
    id: number;
    fullName: string;
  };
  responsible: {
    id: number;
    fullName: string;
  };
}

export type GroupStaffByUser = Pick<GroupStaff, "id" | "groupId" | "personId" | "parentId">;